# Multi path fast filters