This will support the fast filter cache for filters with multiple paths (all leading to `eq`).

Before this commit, the fast filter would only apply to "flat" filters leading to eq. That means `filter: {a: {b: {c: {eq: 5}}}}` would be fast but `filter: {a: {b: {c: {eq: 5}, d: {eq: 6}}}}` (and all varieties) would be slow, take the sift path.

After this commit, any number of filters should take the fast(er) path. It'll not be as fast as a flat path because it needs to do an intersection of all filter conditions (of course), but should still be faster than going through sift.