@pvdz 2020-03-27
Going to run a benchmark to get some numbers and then I'll merge. The Cloud tests failed on some unrelated trunk breakage and I don't expect them to be broken by this.