@pvdz 2020-03-27
Ran some benchmarks;
- benchmarks/markdown_id:
  - 10k pages; old: 131q/s, new: 128 q/s. Total runtime, from 137s to 139s
  - 50k pages; old: 125 q/s, new: 124 q/s
- ifsc (flat filters): 
  - 150k pages; old 2019 q/s, new: 2040 q/s.
- govbook: 
  - 26k pages; old: 1371 q/s, new: 1378 q/s.

- On an internal contentful site, which definitely uses multi-filter queries:
  - 2k pages (many nodes because contentful); old: 21 q/s, after: 20 q/s, total runtime from 209s to 213s
  - 14k pages (1M nodes); 

This seems to be a small regression, even for the site I was trying to improve. That site has other trouble as well so I'm going to have to setup a benchmark to proof this at least fixes the base case it's trying to fix (make sites with multiple filters in a query as fast as those with single filter queries, at scale). So this will have to be Monday.