@pvdz 2020-03-27
Merging since it seems to have the desired effect at scale, even if it's a small regression for non-scale. This will unblock further improvements in this area.